Move config.js to not conflict with config.json
